refactor(eval): split correlation.rs into module directory
Modularize the 1781-line correlation.rs into focused submodules for
maintainability. No behavior changes; identical test results
(366 unit tests pass).

New structure:
  correlation/
    mod.rs       -- module wiring and public re-exports
    types.rs     -- compiled correlation types and conditions
    keys.rs      -- group key extraction
    buffers.rs   -- event buffers and references
    window.rs    -- window state and aggregation helpers
    compiler.rs  -- correlation rule compilation
    tests.rs     -- all unit tests
